+++ CHANGELOG.md
@@ -11,3 +11,7 @@
 - Implemented a WebSocket client-server with websockets/ws library for nodejs environment.
 - Implemented performance test for browser WebTransport client-server application.
 - Implemented performance test for browser WebSockets client-server application.
+
+## Updates
+
+- Added throughput test for both go and nodejs envirnoment.